+++ NEWS.md
@@ -1,5 +1,7 @@
 # filearchy 0.1.0.9000
 
+* Added leaflet html creation. 
+
 * Added `dry_run` to `gdal_tiles()`.